+++ Assets/scripts/MachineGun.cs
@@ -13,7 +13,7 @@ public class MachineGun : MonoBehaviour
 
     public int maxbullet;
     public int curbullet;
-    public int reloadTime;
+    public float reloadTime;
     public float fireRate;
 
     float curfireRate;
